requirements.txt
python-dotenv
pandas
numpy
pymongo
matplotlib
seaborn
certifi
pymongo
pymongo[srv]
pymongo==4.14.1
scikit-learn
dill
pyaml
# -e .
